feat: added function to disable fetching process
Also added functionality which allows for calculating average of the
temperatures in case of multiple measurements from the same device; just
adding it to the sensors list is more than enough.

For example:

sensors = {
    'bmp180': ['altitude', 'pressure', 'seaLevelPressure'],
    'mq135': ['aceton', 'alcohol', 'co', 'co2', 'nh4', 'toulen'],
    'ds18b20': ['temperature'],
    'dht22': ['humidity'],
}

Will take only the DS18B20 sensor's temperature as the main one and
write it right away. However the following:

sensors = {
    'bmp180': [ 'temperature', 'altitude', 'pressure', 'seaLevelPressure'],
    'mq135': ['aceton', 'alcohol', 'co', 'co2', 'nh4', 'toulen'],
    'ds18b20': ['temperature'],
    'dht22': [ 'temperature', 'humidity'],
}

This option can be useful if in possession of a few pretty acurate
sensors. In my case DS18B20 is an average of 8 sensors of this kind, so
taking the average from other, less accurate sensors would only make the
reading less reliable.